Extract Invitation Delete logic to library
